requirements.txt
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
litellm>=1.0.0
pydantic>=2.5.0
pydantic-settings>=2.1.0
redis>=5.0.0
qdrant-client>=1.7.0
sentence-transformers>=2.2.0
vaderSentiment>=3.3.2
python-multipart>=0.0.6
httpx>=0.25.0
structlog>=23.2.0
pyyaml>=6.0
sqlalchemy[asyncio]>=2.0.0
aiosqlite>=0.19.0
prometheus-client>=0.19.0
cryptography>=41.0.0

# Optional: Google Cloud Logging (requires C++ build tools on Windows)
# google-cloud-logging>=3.8.0

# Dev dependencies
pytest>=7.4.0
pytest-asyncio>=0.21.0
pytest-mock>=3.12.0
hypothesis>=6.92.0
fakeredis>=2.20.0
black>=23.12.0
ruff>=0.1.0
